refactor: fix test imports
