+++ src/datastore/offline-dummy-store.ts
@@ -5,6 +5,8 @@ import { PgStore } from './pg-store';
 export const OfflineDummyStore: PgStore = new Proxy(new EventEmitter() as ChainEventEmitter, {
   get(target: any, propKey) {
     // eslint-disable-next-line @typescript-eslint/no-unsafe-return
+    if (propKey === 'eventEmitter') return target;
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-return
     if (propKey in target) return target[propKey];
     return function () {
       throw new Error(
