+++ src/event-replay/helpers.ts
@@ -1,4 +1,4 @@
-import { PgDataStore } from '../datastore/postgres-store';
+import { PgPrimaryStore } from '../datastore/pg-primary-store';
 import { ReverseFileStream } from './reverse-file-stream';
 
 /**
@@ -34,7 +34,7 @@ export async function findTsvBlockHeight(filePath: string): Promise<number> {
  * @param db - Data store
  * @returns Block height
  */
-export async function getDbBlockHeight(db: PgDataStore): Promise<number> {
+export async function getDbBlockHeight(db: PgPrimaryStore): Promise<number> {
   const result = await db.query(async client => {
     return await client.query<{ block_height: number }>(
       `SELECT MAX(block_height) as block_height FROM blocks WHERE canonical = TRUE`
