+++ src/tests/import-genesis-tests.ts
@@ -1,14 +1,15 @@
 import { importV1TokenOfferingData } from '../import-v1';
 import { bitcoinToStacksAddress } from 'stacks-encoding-native-js';
-import { cycleMigrations, PgDataStore, runMigrations } from '../datastore/postgres-store';
+import { PgPrimaryStore } from '../datastore/pg-primary-store';
+import { cycleMigrations, runMigrations } from '../datastore/migrations';
 
 describe('import genesis data tests', () => {
-  let db: PgDataStore;
+  let db: PgPrimaryStore;
 
   beforeEach(async () => {
     process.env.PG_DATABASE = 'postgres';
     await cycleMigrations();
-    db = await PgDataStore.connect({ usageName: 'tests', withNotifier: false });
+    db = await PgPrimaryStore.connect({ usageName: 'tests', withNotifier: false });
   });
 
   test('import token offering data', async () => {
